2 bugfixes

+++ PR/src/include/dat.h
@@ -43,7 +43,7 @@ typedef struct {
 typedef struct {
 	tResourceId        id;
 	tResourceId        palette;
-	unsigned short int size;
+	long int           size;
 	unsigned long int  offset; /* Used internally in dat.c to remember the offset */
 	unsigned char      number; /* Used for level number */
 	char               type;
